refactor: move \`buildTimeoutDiagnostics\` to \`cli-utils\` and reuse in \`gemini.ts\`
Addresses manki-review review comment on PR #678
https://github.com/manki-review/manki/pull/678#discussion_r3209131388
